requirements.txt
pymongo==3.10.0
speedtest-cli==2.1.2
